requirements.txt
BeautifulSoup==3.2.1
Django==1.6.2
argparse==1.2.1
distribute==0.6.24
djangorestframework==2.3.12
ipython==1.2.0
lxml==3.3.1
psycopg2==2.5.2
pysqlite==2.6.3
wsgiref==0.1.2
